add improved version of calculating the max of substring
This Big-O of this version is same as the previous one: O(n) but
reducing steps to find out the max of substring.

In this version, we only need to memorize its last seen position and
move the start of windows to the next position of the last seen
character when the position of the last seen character is larger than
the start of the windows.