add 5_longest_palindromic_substring.cpp
This version of the code was only implemented with dynamic programming.
The solution for this problem with DP is complexity and not a good
choice. So, the code was genereted from claude for the purpose of
learning how to use DP to solve it. Other solutions will be updated in
this file in the future.