fix: elements in nums vector may not be checked
In my point of view, if we can find the initial maximum value in a range
that will not be out of moving windows, we just return the result.
However, we missing the most right of range after the end of board of
always-in windows. This will cause incorrect result.